Find the reason why SimplePriorities doesn't work.
Thread Priorities

The thread priorities available to Java threads on a native threaded JVM
should be treated as hints to the scheduler, especially if the threads are
compute-bound. The number of processors available to a process is dynamic and
unpredictable. Therefore, an attempt to use priorities to schedule execution
on any multi-tasked, multiprocessor system is not likely to succeed.